test: split up cli tests by command; rearrange folders
