+++ test/internals/github-integration.js
@@ -2,8 +2,8 @@
 
 require('dotenv').config()
 const tap = require('tap')
-const github = require('../lib/github')
-const CONFIG = require('./fixtures/config')
+const github = require('../../lib/github')
+const CONFIG = require('../fixtures/config')
 
 tap.test('package.json can be fetched with a valid url', async tap => {
   tap.equal(JSON.stringify(await github.getPackageJson(CONFIG.DEP_ORG, CONFIG.DEP_REPO)), JSON.stringify({ ...CONFIG.PKGJSON, dependencies: { [CONFIG.PKG_NAME_INTEGRATION]: '*' } }))
